Long commit, adds all files required in src, basic rollup build up and running
