--- src/hammer.prefix.js
@@ -1,2 +1,0 @@
-(function(window, document, exportName, undefined) {
-  'use strict';
