+++ src/inputjs/computeDeltaXY.js
@@ -2,6 +2,7 @@
  * Created by arjun on 20/05/16.
  */
 import {INPUT_START,INPUT_END} from './inputConsts'
+
 function computeDeltaXY(session, input) {
     var center = input.center;
     var offset = session.offsetDelta || {};
