+++ src/inputjs/createInputInstance.js
@@ -4,6 +4,10 @@
  */
 import {SUPPORT_POINTER_EVENTS,SUPPORT_ONLY_TOUCH,SUPPORT_TOUCH} from './inputConsts'
 import {inputHandler} from './inputHandler'
+import {PointerEventInput} from '../input/pointerevent';
+import {TouchInput} from '../input/touch';
+import {MouseInput} from '../input/mouse';
+import {TouchMouseInput} from '../input/touchmouse';
 /**
  * create new input type manager
  * called by the Manager constructor
