+++ src/main.js
@@ -1,0 +1,1 @@
+export {Hammer} from './hammer';
