+++ src/recognizerjs/getRecognizerByNameIfManager.js
@@ -1,0 +1,20 @@
+/**
+ * Created by arjun on 25/05/16.
+ */
+
+
+/**
+ * get a recognizer by name if it is bound to a manager
+ * @param {Recognizer|String} otherRecognizer
+ * @param {Recognizer} recognizer
+ * @returns {Recognizer}
+ */
+function getRecognizerByNameIfManager(otherRecognizer, recognizer) {
+    var manager = recognizer.manager;
+    if (manager) {
+        return manager.get(otherRecognizer);
+    }
+    return otherRecognizer;
+}
+
+export {getRecognizerByNameIfManager};
