+++ src/recognizers/attribute.js
@@ -1,3 +1,8 @@
+import {inherit} from '../utilsjs/inherit';
+import {Recognizer} from '../recognizerjs/recognizerConstructor'
+import {STATE_BEGAN,STATE_CHANGED,STATE_CANCELLED,STATE_ENDED,STATE_FAILED} from '../recognizerjs/recognizerConsts';
+import {INPUT_CANCEL,INPUT_END} from '../inputjs/inputConsts';
+
 /**
  * This recognizer is just used as a base for the simple attribute recognizers.
  * @constructor
@@ -58,3 +63,5 @@ inherit(AttrRecognizer, Recognizer, {
         return STATE_FAILED;
     }
 });
+
+export {AttrRecognizer}
