+++ src/touchactionjs/getTouchActionProps.js
@@ -1,0 +1,20 @@
+/*
+ * Created by arjun on 28/05/16.
+ */
+import {NATIVE_TOUCH_ACTION} from './touchactionConsts'
+function getTouchActionProps() {
+    if (!NATIVE_TOUCH_ACTION) {
+        return false;
+    }
+    var touchMap = {};
+    var cssSupports = window.CSS && window.CSS.supports;
+    ['auto', 'manipulation', 'pan-y', 'pan-x', 'pan-x pan-y', 'none'].forEach(function(val) {
+
+        // If css.supports is not supported but there is native touch-action assume it supports
+        // all values. This is the case for IE 10 and 11.
+        touchMap[val] = cssSupports ? window.CSS.supports('touch-action', val) : true;
+    });
+    return touchMap;
+}
+
+export {getTouchActionProps}
