+++ src/utilsjs/addEventListeners.js
@@ -1,7 +1,8 @@
 /**
  * Created by arjun on 20/05/16.
  */
-import {each} from './each'
+import {each} from './each';
+import {splitStr} from './splitStr';
 /**
  * addEventListener with multiple events at once
  * @param {EventTarget} target
