+++ src/utilsjs/ifUndefined.js
@@ -1,0 +1,15 @@
+/**
+ * Created by arjun on 25/05/16.
+ */
+
+/**
+ * use the val2 when val1 is undefined
+ * @param {*} val1
+ * @param {*} val2
+ * @returns {*}
+ */
+function ifUndefined(val1, val2) {
+    return (val1 === undefined) ? val2 : val1;
+}
+
+export {ifUndefined};
