+++ src/utilsjs/setTimeoutContext.js
@@ -1,7 +1,7 @@
 /**
  * Created by arjun on 19/05/16.
  */
-
+import {bindFn} from './bindFn';
 /**
  * set a timeout with a given scope
  * @param {Function} fn
